+++ client/app/left/topics/dropdown/left.folder.dropdown.layer.controller.js
@@ -1,6 +1,6 @@
 /**
  * @fileoverview 센터페널 메세지중 단순 텍스트일때만 관리하는 컨트롤러
- * @author JiHoon Kim <jihoonk@tosslab.com>
+ * @author Young Park <young.park@tosslab.com>
  */
 (function() {
   'use strict';
@@ -24,10 +24,16 @@
     function _init() {
     }
 
+    /**
+     * 폴더에서 제거한다.
+     */
     function remove() {
       TopicFolderModel.remove($scope.folder.id);
     }
 
+    /**
+     *
+     */
     function rename() {
       jndPubSub.pub('topic-folder:rename', $scope.folder.id);
     }