+++ client/app/util/service/jnd.util.service.js
@@ -8,9 +8,15 @@
     .module('jandiApp')
     .service('JndUtil', JndUtil);
 
-  function JndUtil() {
+  function JndUtil(Dialog) {
     this.safeApply = safeApply;
+    this.alertUnknownError = alertUnknownError;
 
+    /**
+     * angular 의 $apply 를 안전하게 수행한다.
+     * @param {object} scope
+     * @param {function} fn
+     */
     function safeApply(scope, fn) {
       if (scope) {
         fn = _.isFunction(fn) ? fn : function () {
@@ -21,6 +27,20 @@
           fn();
         }
       }
+    }
+
+    /**
+     * 알수 없는 오류에 대한 alert 을 노출한다.
+     * @param response
+     */
+    function alertUnknownError(response) {
+      response = _.extend({
+        code: -1,
+        msg: 'unknown error'
+      }, response);
+      Dialog.alert({
+        body: 'error :' + response.code + '\n' + response.msg
+      });
     }
   }
 })();