+++ CHANGELOG.md
@@ -1,3 +1,119 @@
+<a name"1.1.5"></a>
+### 1.1.5 (2015-09-23)
+
+
+#### Bug Fixes
+
+* **404:** 디자인 pull request 적용   - https://github.com/tosslab/web_client/pull/218/files ([8976ecda](https://github.com/tosslab/web_client.git/commit/8976ecda))
+* **L10N:**
+  * L10N 수정 ([5279f0c5](https://github.com/tosslab/web_client.git/commit/5279f0c5))
+  * 로그인 페이지 가입하기 L10N 적용 ([c4ad65bb](https://github.com/tosslab/web_client.git/commit/c4ad65bb))
+* **center:**
+  * loading 휠 관련 수정 ([1fd00f60](https://github.com/tosslab/web_client.git/commit/1fd00f60))
+  * 공유된 대화방 이동 L10N 수정 ([089bf58c](https://github.com/tosslab/web_client.git/commit/089bf58c))
+  * 서버 응답이 500ms 보다 늦어질 경우 center loading wheel 을 노출한다. ([13116e6e](https://github.com/tosslab/web_client.git/commit/13116e6e))
+  *  CSS Pull request 적용   - https://github.com/tosslab/web_client/pull/202/files ([311e9fcb](https://github.com/tosslab/web_client.git/commit/311e9fcb))
+* **config:**
+  * package.json 콘피그 오류 수정 ([1a454b7d](https://github.com/tosslab/web_client.git/commit/1a454b7d))
+  * config.js 자동으로 생성되므로 git에서 제거 ([9831153c](https://github.com/tosslab/web_client.git/commit/9831153c))
+  * git ignore 에 config 도 추가 ([24269d94](https://github.com/tosslab/web_client.git/commit/24269d94))
+  * git ignore 에 local.team.json 제거 ([a0bfb841](https://github.com/tosslab/web_client.git/commit/a0bfb841))
+  * tosslab 이 기본 팀이 되도록 수정 ([3d58967b](https://github.com/tosslab/web_client.git/commit/3d58967b))
+  * local 작업에서 config 파일로 team prefix 수정 가능하도록 추가 ([7250d6b3](https://github.com/tosslab/web_client.git/commit/7250d6b3))
+* **confilct:** resolve conflict with develop ([ae09b3ae](https://github.com/tosslab/web_client.git/commit/ae09b3ae))
+* **dialog:**
+  * 비공개 토픽 나가기시 Dialog 사용하도록 수정 ([99955b1a](https://github.com/tosslab/web_client.git/commit/99955b1a))
+  * 비공개 토픽 나가기시 Dialog 사용하도록 수정 ([04ff98f9](https://github.com/tosslab/web_client.git/commit/04ff98f9))
+* **dm.file.share:**
+  * Update center when file is uploaded from me. ([9c7637ea](https://github.com/tosslab/web_client.git/commit/9c7637ea))
+  * Update center and lef on first file share event.   -#JND-3458 #resolve ([8b597014](https://github.com/tosslab/web_client.git/commit/8b597014))
+  * Update chat list when file is shared to dm.   - #JND-3458 #resolve ([26a90fe9](https://github.com/tosslab/web_client.git/commit/26a90fe9))
+* **favicon:** Change favicon image assets. ([a4e811ad](https://github.com/tosslab/web_client.git/commit/a4e811ad))
+* **file.upload:**
+  * upload api v2 사용하도록 수정 ([a353a80a](https://github.com/tosslab/web_client.git/commit/a353a80a))
+  * file upload시 작성된 message가 comment로 들어갈 수 있도록 수정 ([3ff3d046](https://github.com/tosslab/web_client.git/commit/3ff3d046))
+* **grunt:** fix version release script ([00a915e6](https://github.com/tosslab/web_client.git/commit/00a915e6))
+* **left:**
+  * 리뷰용 주석 추가 ([f2030a36](https://github.com/tosslab/web_client.git/commit/f2030a36))
+  * 폴더명 변경 시 input 에서 텍스트 드래그 불가한 현상 수정 ([debb4ed1](https://github.com/tosslab/web_client.git/commit/debb4ed1))
+  * 검색이나, push로 진입한 토픽의 폴더명이 Bold 표시되지 않음 문제 해결   - 더블클릭 시 디폴트 토픽으로 이동하는 문제 해결   - J ([8f6ba2c7](https://github.com/tosslab/web_client.git/commit/8f6ba2c7))
+  * 폴더를 접을 때 [ … ] 아이콘 위에 빨간 unread badge가 오버랩 되었다가 사라짐 현상 해결   - JND-3515 #resolve ([fd3f67f9](https://github.com/tosslab/web_client.git/commit/fd3f67f9))
+  * 코드리뷰 준비   - 주석 추가 및 코드 정리   - safeApply 적용 ([fb075e9e](https://github.com/tosslab/web_client.git/commit/fb075e9e))
+  * go unread below 수정 ([30a47de6](https://github.com/tosslab/web_client.git/commit/30a47de6))
+* **mention:**
+  * mention 작성시 "@all" 출력되지 않는 버그 수정 ([f07adabc](https://github.com/tosslab/web_client.git/commit/f07adabc))
+  * mention 작성시 "@all" 출력되는 text 수정 ([3945cf92](https://github.com/tosslab/web_client.git/commit/3945cf92))
+* **modals:** 버그 및 기능 변경   - safari에서 scrolling 가속 적용되도록 수정   - firefox에서 tab 변경시 item이 생성되지 않 ([ef58902e](https://github.com/tosslab/web_client.git/commit/ef58902e))
+* **readme:**
+  * 내용 추가 ([da6c1994](https://github.com/tosslab/web_client.git/commit/da6c1994))
+  * 내용 추가 ([44479759](https://github.com/tosslab/web_client.git/commit/44479759))
+* **right:**
+  * integration file download ([59bc0ddf](https://github.com/tosslab/web_client.git/commit/59bc0ddf))
+  * right panel에서 사용하는 message card가 disable member를 표현하도록 수정 ([1fc65c29](https://github.com/tosslab/web_client.git/commit/1fc65c29))
+* **signin:** 주석 추가 ([ce8d8fbf](https://github.com/tosslab/web_client.git/commit/ce8d8fbf))
+
+
+<a name"1.1.5"></a>
+### 1.1.5 (2015-09-23)
+
+
+#### Bug Fixes
+
+* **404:** 디자인 pull request 적용   - https://github.com/tosslab/web_client/pull/218/files ([8976ecda](https://github.com/tosslab/web_client.git/commit/8976ecda))
+* **L10N:**
+  * L10N 수정 ([5279f0c5](https://github.com/tosslab/web_client.git/commit/5279f0c5))
+  * 로그인 페이지 가입하기 L10N 적용 ([c4ad65bb](https://github.com/tosslab/web_client.git/commit/c4ad65bb))
+* **center:**
+  * loading 휠 관련 수정 ([1fd00f60](https://github.com/tosslab/web_client.git/commit/1fd00f60))
+  * 공유된 대화방 이동 L10N 수정 ([089bf58c](https://github.com/tosslab/web_client.git/commit/089bf58c))
+  * 서버 응답이 500ms 보다 늦어질 경우 center loading wheel 을 노출한다. ([13116e6e](https://github.com/tosslab/web_client.git/commit/13116e6e))
+  *  CSS Pull request 적용   - https://github.com/tosslab/web_client/pull/202/files ([311e9fcb](https://github.com/tosslab/web_client.git/commit/311e9fcb))
+* **config:**
+  * package.json 콘피그 오류 수정 ([1a454b7d](https://github.com/tosslab/web_client.git/commit/1a454b7d))
+  * config.js 자동으로 생성되므로 git에서 제거 ([9831153c](https://github.com/tosslab/web_client.git/commit/9831153c))
+  * git ignore 에 config 도 추가 ([24269d94](https://github.com/tosslab/web_client.git/commit/24269d94))
+  * git ignore 에 local.team.json 제거 ([a0bfb841](https://github.com/tosslab/web_client.git/commit/a0bfb841))
+  * tosslab 이 기본 팀이 되도록 수정 ([3d58967b](https://github.com/tosslab/web_client.git/commit/3d58967b))
+  * local 작업에서 config 파일로 team prefix 수정 가능하도록 추가 ([7250d6b3](https://github.com/tosslab/web_client.git/commit/7250d6b3))
+* **confilct:** resolve conflict with develop ([ae09b3ae](https://github.com/tosslab/web_client.git/commit/ae09b3ae))
+* **dialog:**
+  * 비공개 토픽 나가기시 Dialog 사용하도록 수정 ([99955b1a](https://github.com/tosslab/web_client.git/commit/99955b1a))
+  * 비공개 토픽 나가기시 Dialog 사용하도록 수정 ([04ff98f9](https://github.com/tosslab/web_client.git/commit/04ff98f9))
+* **dm.file.share:**
+  * Update center when file is uploaded from me. ([9c7637ea](https://github.com/tosslab/web_client.git/commit/9c7637ea))
+  * Update center and lef on first file share event.   -#JND-3458 #resolve ([8b597014](https://github.com/tosslab/web_client.git/commit/8b597014))
+  * Update chat list when file is shared to dm.   - #JND-3458 #resolve ([26a90fe9](https://github.com/tosslab/web_client.git/commit/26a90fe9))
+* **favicon:** Change favicon image assets. ([a4e811ad](https://github.com/tosslab/web_client.git/commit/a4e811ad))
+* **file.upload:**
+  * upload api v2 사용하도록 수정 ([a353a80a](https://github.com/tosslab/web_client.git/commit/a353a80a))
+  * file upload시 작성된 message가 comment로 들어갈 수 있도록 수정 ([3ff3d046](https://github.com/tosslab/web_client.git/commit/3ff3d046))
+* **grunt:** fix version release script ([00a915e6](https://github.com/tosslab/web_client.git/commit/00a915e6))
+* **left:**
+  * 리뷰용 주석 추가 ([f2030a36](https://github.com/tosslab/web_client.git/commit/f2030a36))
+  * 폴더명 변경 시 input 에서 텍스트 드래그 불가한 현상 수정 ([debb4ed1](https://github.com/tosslab/web_client.git/commit/debb4ed1))
+  * 검색이나, push로 진입한 토픽의 폴더명이 Bold 표시되지 않음 문제 해결   - 더블클릭 시 디폴트 토픽으로 이동하는 문제 해결   - J ([8f6ba2c7](https://github.com/tosslab/web_client.git/commit/8f6ba2c7))
+  * 폴더를 접을 때 [ … ] 아이콘 위에 빨간 unread badge가 오버랩 되었다가 사라짐 현상 해결   - JND-3515 #resolve ([fd3f67f9](https://github.com/tosslab/web_client.git/commit/fd3f67f9))
+  * 코드리뷰 준비   - 주석 추가 및 코드 정리   - safeApply 적용 ([fb075e9e](https://github.com/tosslab/web_client.git/commit/fb075e9e))
+  * go unread below 수정 ([30a47de6](https://github.com/tosslab/web_client.git/commit/30a47de6))
+* **mention:**
+  * mention 작성시 "@all" 출력되지 않는 버그 수정 ([f07adabc](https://github.com/tosslab/web_client.git/commit/f07adabc))
+  * mention 작성시 "@all" 출력되는 text 수정 ([3945cf92](https://github.com/tosslab/web_client.git/commit/3945cf92))
+* **modals:** 버그 및 기능 변경   - safari에서 scrolling 가속 적용되도록 수정   - firefox에서 tab 변경시 item이 생성되지 않 ([ef58902e](https://github.com/tosslab/web_client.git/commit/ef58902e))
+* **readme:** 내용 추가 ([44479759](https://github.com/tosslab/web_client.git/commit/44479759))
+* **right:**
+  * integration file download ([59bc0ddf](https://github.com/tosslab/web_client.git/commit/59bc0ddf))
+  * right panel에서 사용하는 message card가 disable member를 표현하도록 수정 ([1fc65c29](https://github.com/tosslab/web_client.git/commit/1fc65c29))
+* **signin:** 주석 추가 ([ce8d8fbf](https://github.com/tosslab/web_client.git/commit/ce8d8fbf))
+
+
+<a name"1.1.3"></a>
+### 1.1.3 (2015-09-17)
+
+
+<a name"1.1.1"></a>
+### 1.1.1 (2015-09-17)
+
+
 <a name"1.1.0"></a>
 ## 1.1.0 (2015-09-16)